requirements.txt
Django==1.9.5
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
Pillow==3.1.0
djangorestframework==3.3.3
Markdown=2.6.6
django-filter==0.13.0
djangorestframework-jwt==1.8.0